tgidemo fixes
COLOR_BACK and COLOR_FORE are used as palette indices, not color values.
The previous code defining them to COLOR_BLACK and COLOR_WHITE only
works on platforms where those values happen to be 0 and 1.

Move keypress detection to the inner loop of DoCircles, so as to match
other routines.